requirements.txt
# Camping Deals - Dependencias Python
# 100% gratuitas y open source

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0
lxml>=5.0.0

# Base de datos (Supabase es opcional)
supabase>=2.3.0

# Email Marketing
resend>=0.7.0

# Bot Telegram
python-telegram-bot>=20.7

# Analytics
gspread>=5.12.0
google-auth>=2.25.0

# Utilidades
python-dotenv>=1.0.0
schedule>=1.2.0
